+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Raytracing/HDRaytracingManager.cs
@@ -9,6 +9,14 @@ using UnityEditor;
 
 namespace UnityEngine.Rendering.HighDefinition
 {
+    [GenerateHLSL]
+    internal enum RayTracingRendererFlag
+    {
+        Opaque = 0x01,
+        Transparent = 0x02,
+        CastShadow = 0x04
+    }
+
 #if ENABLE_RAYTRACING
     class HDRayTracingLights
     {
@@ -580,7 +588,7 @@ namespace UnityEngine.Rendering.HighDefinition
                             uint instanceFlag = 0x00;
 
                             // Incorporate the shadow casting flag
-                            instanceFlag |= ((currentRenderer.shadowCastingMode == ShadowCastingMode.On) ? (uint)(1 << 2) : 0x00);
+                            instanceFlag |= ((currentRenderer.shadowCastingMode == ShadowCastingMode.On) ? (uint)(RayTracingRendererFlag.CastShadow) : 0x00);
 
                             for (int meshIdx = 0; meshIdx < numSubMeshes; ++meshIdx)
                             {
@@ -598,7 +606,7 @@ namespace UnityEngine.Rendering.HighDefinition
                                     && HDRenderQueue.k_RenderQueue_AllTransparentRaytracing.upperBound >= currentMaterial.renderQueue);
 
                                     // Propagate the right mask
-                                    instanceFlag |= materialIsTransparent ? (uint)(1 << 1) : (uint)(1 << 0);
+                                    instanceFlag |= materialIsTransparent ? (uint)(RayTracingRendererFlag.Transparent) : (uint)(RayTracingRendererFlag.Opaque);
 
                                     // Is the material alpha tested?
                                     subMeshCutoffArray[meshIdx] = currentMaterial.IsKeywordEnabled("_ALPHATEST_ON")